auth-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.1
python-multipart==0.0.6
httpx==0.25.2
slowapi==0.1.9
psycopg2-binary==2.9.9


# Shared dependencies
# Shared library dependencies
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-jose[cryptography]>=3.3.0
slowapi>=0.1.9
httpx>=0.24.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

